+++ spec/javascripts/module.stop.spec.js
@@ -55,4 +55,14 @@ describe("module stop", function(){
     });
   });
 
+  describe("when stopping a sub-module from the parent module, by name", function(){
+    it("should stop the specified sub-module", function(){
+      throw "not yet implemented";
+    });
+
+    it("should not stop the parent module", function(){
+      throw "not yet implemented";
+    });
+  });
+
 });